sc3: Подкручивание цвета панелей.
cordova: Менять цвет системного статус-бара согласно цвета фона панели.

+++ src1/client/geo/common-maps/src/main/scala/io/suggest/maps/r/LGeoMapR.scala
@@ -43,7 +43,7 @@ object LGeoMapR {
       },
 
       onZoomEndF = ReactCommonUtil.cbFun1ToJsCb { event: Event =>
-        val newZoom = event.target.asInstanceOf[LMap].getZoom()
+        val newZoom = event.target.asInstanceOf[LMap].getZoom().toInt
         ReactDiodeUtil.dispatchOnProxyScopeCB( $, MapZoomEnd(newZoom) )
       },
